% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\docType{data}
\name{spotted_year_fit}
\alias{spotted_year_fit}
\title{Fitted glmmLDTS Model Object for Spotted Seals - With Year}
\format{
A glmmLDTS object with 21 elements:
\describe{
\item{dataset}{data used for model fit}
\item{WARNINGS}{}
\item{fixed.formula}{class \code{formula} for fixed effects}
\item{random.formula}{class \code{formula} for random effects}
\item{sample.size}{number of records in the data set}
\item{timecol}{column that specifies the time vector}
\item{trialscol}{}
\item{group.vec}{column identifying groups}
\item{ridge.reg}{}
\item{lambda}{}
\item{start.time}{start time for model fit}
\item{end.time}{end time for model fit}
\item{R.cov.paramaters}{}
\item{G.cov.parameters}{}
\item{fixed.effects}{fixed effects from the model}
\item{random.effects}{random effects from the model}
\item{typeIII.hypoth}{}
\item{fit.table}{}
\item{covb}{}
\item{outer.iterations}{}
\item{inner.iterations2}{}
}
}
\usage{
spotted_year_fit
}
\description{
A fitted glmmLDTS model object for spotted seals that includes
a year covariate in the model formula
}
\keyword{datasets}
